* Return the straight size_t rather than implicitly convert to int

+++ src/common/SimpleCollection.h
@@ -332,7 +332,7 @@ public:
 	 * @return int
 	 * 		The collections size
 	 */
-	virtual int getSize() const {
+	virtual size_t getSize() const {
 		return objects.size();
 	}
